+++ test/useSwipe.spec.js
@@ -1,0 +1,47 @@
+import { cleanup, renderHook } from '@testing-library/react-hooks';
+import useSwipe from '../dist/useSwipe';
+import useHorizontalSwipe from '../src/useHorizontalSwipe';
+import useVerticalSwipe from '../src/useVerticalSwipe';
+
+describe('useSwipe', () => {
+  beforeEach(cleanup);
+
+  it('should be a function', () => {
+    expect(useSwipe).to.be.a('function');
+  });
+
+  it('should return the swipe state', () => {
+    const { result } = renderHook(() => useSwipe());
+
+    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alpha', 'count');
+  });
+});
+
+describe('useHorizontalSwipe', () => {
+  beforeEach(cleanup);
+
+  it('should be a function', () => {
+    expect(useHorizontalSwipe).to.be.a('function');
+  });
+
+  it('should return the swipe state', () => {
+    const { result } = renderHook(() => useHorizontalSwipe());
+
+    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alpha', 'count');
+  });
+});
+
+
+describe('useVerticalSwipe', () => {
+  beforeEach(cleanup);
+
+  it('should be a function', () => {
+    expect(useVerticalSwipe).to.be.a('function');
+  });
+
+  it('should return the swipe state', () => {
+    const { result } = renderHook(() => useVerticalSwipe());
+
+    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alpha', 'count');
+  });
+});